add user page, fix login and authorization issue

+++ src/Controller/ProductController.php
@@ -3,8 +3,10 @@
 namespace App\Controller;
 
 use App\Entity\Product;
+use App\Service\AuthenticationChecker;
 use App\Service\ProductListService;
 use App\Service\ProductService;
+use App\Traits\AuthenticationCheckerTrait;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -13,31 +15,45 @@ use Symfony\Component\Routing\Annotation\Route;
 
 class ProductController extends AbstractController
 {
+    use AuthenticationCheckerTrait;
+
     private ProductListService $productListService;
 
     private ProductService $productService;
 
     private EntityManagerInterface $entityManager;
 
+    private AuthenticationChecker $authenticationChecker;
+
     public function __construct(
         ProductListService $productListService,
         ProductService $productService,
-        EntityManagerInterface $entityManager
+        EntityManagerInterface $entityManager,
+        AuthenticationChecker $authenticationChecker
     ) {
         $this->productListService = $productListService;
         $this->productService = $productService;
         $this->entityManager = $entityManager;
+        $this->authenticationChecker = $authenticationChecker;
     }
 
     #[Route('/api/product/list', name: 'api_product_list')]
     public function listAction(): JsonResponse
     {
+        if (null !== $invalidAuthentication = $this->isAuthenticationInvalid()) {
+            return $invalidAuthentication;
+        }
+
         return new JsonResponse($this->productListService->getAllProducts());
     }
 
     #[Route('/api/product/edit/{id}', name: 'api_product_edit')]
     public function editAction(Request $request): JsonResponse
     {
+        if (null !== $invalidAuthentication = $this->isAuthenticationInvalid()) {
+            return $invalidAuthentication;
+        }
+
         $productId = (int) $request->get('id');
         return new JsonResponse($this->productService->getProductById($productId));
     }
@@ -45,6 +61,10 @@ class ProductController extends AbstractController
     #[Route('/api/product/new', name: 'api_product_new', methods: ["POST"])]
     public function newAction(Request $request): JsonResponse
     {
+        if (null !== $invalidAuthentication = $this->isAuthenticationInvalid()) {
+            return $invalidAuthentication;
+        }
+
         $product = $this->getProduct(json_decode($request->getContent(), true));
         try {
             $this->entityManager->persist($product);
@@ -59,6 +79,10 @@ class ProductController extends AbstractController
     #[Route('/api/product/delete/{id}', name: 'api_product_delete', methods: ["DELETE"])]
     public function deleteAction(Request $request): JsonResponse
     {
+        if (null !== $invalidAuthentication = $this->isAuthenticationInvalid()) {
+            return $invalidAuthentication;
+        }
+
         try {
             $this->productService->removeProductById($request->get('id'));
         } catch (\Exception) {
